gcf-function/requirements.txt
tensorflow>=2.4.0
numpy>=1.19.4
Pillow>=8.0.1
google-cloud==0.34.0
Flask==1.1.2
